WL#8351: Use statistics to determine #objects in Ndb_free_list_t
The utility 'class NdbStatistics' is added which allow
incremental statistics to be calculated over a sliding windows
of statistics samples.

A member of type 'class NdbStatistics' is then added to
each instance of 'template<class T> Ndb_free_list_t'
and used to sample the max number of objects in use
each time a peak is reached.

An upper 95% percentile of objects in use is then
calculated for the statistics window, and used to determine
if a released objects should be stored in the free_list
for later reuse, or returned to the heap manager (deleted)

This should avoid that temporary peaks in NdbApi object
allocated, cause a permanently increase in the memory
footprint of the API client / mysqld.