Bug#23061746 EXPORT MY_MAKE_SCRAMBLED_PASSWORD() IN LIBMYSQLCLIENT
Problem: The list of symbols exported by libmysqlclient.so was
restricted by the fix for bug #20476596. The list of symbols was
restricted to two classes of functions:

1. The public documented API
2. Undocumented symbols that we know are used

Pure-FTPd depends on a symbol that was not included:
my_make_scrambled_password. If that symbols is not found, Pure-FTPd
will fall back to make_scrambled_password, which is also hidden in
libmysqlclient.

Fix: Export my_make_scrambled_password in libmysqlclient.so.

The symbol will be removed again in a future version.

Since we're adding a symbol, bump the minor ABI version number from 2
to 3.