Bug #22017120 CRASH OF MYSQLPUMP
Problem: When dump process is in progress and if one of the underlying
thread causes an error then we call Program::error() which in turn calls
exit(). As part of exit() there is a cleanup done which calls destructor
for mysql_chain_element, mysql_crawler etc. If during this cleanup process
if there is a context switch to other thread, then there is a possibility
of crash if this thread accesses deleted memory.

Fix: Fix is to have a variable store the error code in main program class
and pass this to abstract_crawler and object_queue. In
Object_queue::queue_thread() and Abstract_crawler::wait_for_tasks_completion()
there is a periodic check for this variable for any errors. If there is an
error then simply return from these function calls.