Bug#21352937 - REDUCE LOG_SYS->MUTEX CONTENTION BY ALLOWING CONCURRENT MTR COMMIT AND LOG WRITE
All mtr commits would have to wait for writing/flushing redo logs to disk
in log_write_up_to(). The wait is unnecessary. In this patch, we introduced
one more log buffer and one mutex.

We always write logs to one log buffer, keeping the other idle. Once we
need to flush it, we switch these two buffers, so log_write_up_to() would
take over the to be flushed buffer and do the flush job, once the flushing
finishes, this buffer becomes the idle one. At the mean time of flushing,
all concurrent mtr commits would see the empty new buffer and write logs
there, even after the flushing, logs are still writing to this buffer
before it needs to be flushed.

Let's say we have two adjacent log buffer A and B, at first, all logs are
written to A, keeping B as empty. Once we want to flush logs in A, we will
switch A and B, which is under the protection of current log_sys->mutex.
Then all logs are now written to B, while A is getting flushed and becomes
empty and idle. When B is needed to be flushed, we now switch A and B again,
new commits will come to A, and flushing of B won't stop the concurrent
mtr commits. And so on.

The new introduced mutex is mainly used to protect the process of
log_write_up_to().

The original patch was provided by Weixiang Zhai.

RB: 11142
Reviewed-by: Debarun Banerjee <debarun.banerjee@oracle.com>