Bug #21616496: CREATE USER ACCEPTS PLUGIN AND PASSWORD HASH,                BUT IGNORES THE PASSWORD HASH
Description: Creating a user with 'mysql_native_password'/
'mysql_old_password' as plugin and using "IDENTIFIED WITH
plugin AS 'hash_string'" as syntax, creates one without
a password.

Analysis: In 5.5 and 5.6, the "AS" part in "IDENTIFIED WITH
... AS <hash string>" serves a different purpose and not for
storing the password. According to the documentation, the
<hash string> is stored into the "authentication_string"
column of the mysql.user table. And the password for the
native authentication plugin is never read from this column
instead its read from the "password" column.

Fix: The bug is reclassified as a password validation plugin
bug. The fix is to throw an error, "ERROR 1819 (HY000): Your
password does not satisfy the current policy requirements",
when a user is created with 'mysql_native_password'/
'mysql_old_password' as plugin and using "IDENTIFIED WITH
plugin AS 'hash_string'" as syntax. The error is thrown only
if the "validate_password" plugin is installed.