Bug#22899690  FAILING ASSERT: FIL_SPACE_GET(TABLE->SPACE) != __NULL IN ROW0QUIESCE.CC LINE 724
Issue:
======
When we discard a tablespace for a table and we do a FLUSH TABLE on
the same table, we hit a FIL_SPACE_GET(TABLE->SPACE) !=__NULL assert
in row_quiesce_table_start().

Fix:
====
After discarding the tablespace, flushing is not meaningful. So we
check and return an error if the tablespace has been discarded, even
before we reach the assert.

RB: 12141
Reviewed-by: Debarun Banerjee <debarun.banerjee@oracle.com>