Bug#23097750 DO NOT BUILD BUNDLED PROTOBUF IF IT IS NOT NEEDED
Currently Protobuf is used only by the X plugin.
Building of X plugin can be disabled with cmake -DWITH_RAPID=0
Still, we will build protobuf libraries and executables.

This is a waste of resources on most platform, but a showstopper when building
with SolarisStudio, since the bundled protobuf-2.6.0 does not build at all.

Solution: Tag protobuf targets with
SET_PROPERTY(TARGET xxx PROPERTY EXCLUDE_FROM_ALL TRUE)
so that building of protobuf is skipped if we do cmake -DWITH_RAPID=0