BUG #22608247 RESULT CONTENT MISMATCH IN BINLOG.BINLOG_MYSQLBINLOG_ROW AND FRIENDS
Problem:
--------
This is a known bug in Visual Studio 2015 C Runtime (Connect#1902345)
https://connect.microsoft.com/VisualStudio/feedback/details/1902345

The bug in _read is as follows:  If...

1.       you are reading from a text mode pipe,
2.       you call _read to read N bytes,
3.       _read successfully reads N bytes, and
4.       the last byte read is a carriage return (CR) character,

then the _read function will complete the read successfully but will
return N-1 instead of N.  The CR or LF character at the end of the result
buffer is not counted in the return value.

The bug is fundamentally timing-sensitive because whether _read can
successfully read N bytes from the pipe depends on how much data has been
written to the pipe.  Changing the buffer size or changing when the buffer
is flushed may reduce the likelihood of the problem, but it won't
necessarily work around the problem in 100% of cases.

Fix:
----
The workaround is to use a binary pipe and do text mode CRLF => LF translation
manually on the reader side.

Note that this workaround should no longer be necessary when the next
update to the Universal CRT ships, which is likely to occur around the same
time as the Windows 10 Anniversary Update this summer (2016).

Reviewed-By: Bjorn Munch <bjorn.munch@oracle.com>
RB: 12384