Bug#23110596: RPL.RPL_MULTI_SOURCE_PERFSCHEMA FAILS ON               PB2 WITH RESULT CONTENT MISMATCH
Issue:
------
The master sends heartbeats to the slave only if there
are no unsent events in the binary log file for a
period longer than a given interval.
The number of heartbeats sent will therefore depend on
the machine. This is why rpl.rpl_multi_source_perfschema
was failing sporadically with result content mismatch.

Fix:
----
The fix is to mask the values of COUNT_RECEIVED_HEARTBEATS
and LAST_HEARTBEAT_TIMESTAMP in the result.

Reviewed-by: Sujatha Sivakumar <sujatha.sivakumar@oracle.com>
RB: 12403