BUG#23121777 - ADD SERVER SIDE SUPPORT TO SKIP GTIDS ON GROUP REPLICATION
In order to add this functionality to add support to skip GTID for group
replication we need to have changed in the server side.

We need to be sure that the server operating state when we try handle
the connections, should be SERVER_OPERATING. This requires moving the
before_handle_connection hook to be moved to a new location.