Bug#22255698 INAPPROPRIATE LIST ITERATOR USAGE IN 'INIT_GTID_SETS()'
Problem:

When issuing a CHANGE MASTER for a channel that doesn't exist yet, the
init_gtid_set() function is called to compute the retrieved GTID set of
a replication channel that doesn't have relay log files.

In this case, when going to parse the relay log for incomplete
transactions, the function uses a reverse list iterator for the relay
log file names, incrementing it when already pointing at the being of
the list and decrementing it when pointing at the end.

Also, the reverse iterator pointing at begin/end was used to search the
file name on a list.

Fix:

Only do the parsing for incomplete transactions if there are files in
the relay log for the channel.

Added also some asserts to verify the proper behavior.