Bug #22343910: SELECT DISTINCT NOT RETURNING DISTINCT ROWS                IN 5.7
ISSUE:
------
This issue occurs under the following conditions:
1) A column with a default value stores a null.
2) A distinct or group-by is used in the query and the
   column with the null value is part of the select-list.
3) The temporary table is created on INNODB.

Root-cause
----------
While retrieving the values for the temp table, INNODB
returns the default value of the column
(See row_sel_store_mysql_field_func). The executor,
ignoring the null flag, assumes that the column actually
contains data and treats it as a distinct row.

This problem doesn't occur when MYISAM is used for creating
temporary table because, the default value is not returned
while adding a row of the temporary table.

This problem doesn't occur when HEAP memory is used since
it takes the null flag into account while evaluating keys
(See hp_rec_key_cmp).

This problem occurs only in 5.7+ since INNODB temp tables
are allowed on these versions.


SOLUTION:
---------
While evaluating temporary table values, the executor
shouldn't ignore the null flag of a column.

Also changing the names of the variables src_len and
dst_len. The names can be misleading, since this is a
comparision function.