BUG#23067038 ASSERTION FAILURE: BUF0BUF.CC:2861:BUF_PAGE_IN_FILE(BPAGE)              LEADS TO CORRUPT DATA
It's a regression of wl#8423 InnoDB: Split the buffer pool mutex,
in which we removed block mutex protection for buf_fix_count.

The assertion happens when one thread fixed a dirty block but the
other thread flushed the block and moved the page from LRU list to
free list, because it saw buf_fix_count is 0, other than 1.

The solution is holding block mutex when buf fix and unfix.

Reviewed-by: Debarun Banerjee <debarun.banerjee@oracle.com>
RB: 12456