Bug#22102456 ENABLING GTID AT STARTUP RESULTS IN UNNECESSARY GTID SETS
Problem:

When a server is initialized with GTID enabled, it creates a large and
unnecessary GTID set.

Fix:

Created an infrastructure to differentiate if a query to be processed
during server initialization is a compiled one or if it belongs to a
file.

Disabled binary logging session option before starting to apply
compiled initialize statements and re-enabled it, if needed, when
starting to apply file statement or when the initialize has finished.