WL#9269: X Protocol StmtExecute new admin command ensure_collection
Description
===========
Additional command understood in the xplugin namespace:
ensure_collection. 
Creates a table for document storage (similar to create_collection).
Document tables are created with the format described in x-protocol
documentation. They are to be treated as a opaque document bucket and
is not intended to have its structure changed by the user.
If document table is already exist then no action is being performed
and no error is being raised.


Approved by Lalit Choudhary <lalit.choudhary@oracle.com>