WL#7743 Work-in-progress.
Ensure that reopening of partitioning tables doesn't fail after
failed DDL. The problem is that Item_field::fix_fields() which
might be called when opening table with partitioning expression
will always fail if there is error status in current diagnostics
area. We circumvent this problem by installing temporary DA for
the duration of table re-open.