Bug#23074801: Delete from joined tables with where using derived table               fails with error 1093
This is a regression in 5.7 and occurs because subqueries are now
resolved later than they used to be in 5.6. When the unique_table()
test is performed Sql_cmd_delete_multi::mysql_multi_delete_prepare(),
the subqueries are not yet resolved, and hence the query blocks have
not yet have the flag exclude_from_table_unique_test set.

The problem is fixed by delaying the uniqueness check to the function
Query_result_delete::prepare(), which is called after the subqueries
have been resolved.