BUG#23090538 : MTR KILLS THE SESSION IN 60SEC WHEN USING --MANUAL-GDB OPTION
Issue :
=======
The if-elsif ladder check for diffrent debugger option was broken.
When --manual-gdb option is passed. MTR doesn't wait for the user
to start the server and makes several attempts to connect to server.
If the server is not started within the time(500 attempts) then,
connection timeout error occurs.

Fix :
=====
Fix the broken if-elsif ladder whcih makes MTR to call sleep_until_file_created()
subroutine and wait for the server to start.

Reviewed-by: Sayantan Dutta <sayantan.dutta@oracle.com>
Reviewed-by: Gopal Shankar <gopal.shankar@oracle.com>
RB: 12416