Bug #22986823 HIGH CPU USAGE FOR ~NDB
Patch fixes a problem with 'unfairness' in how
TransporterRegistry::performReceive() receives and delivers
signals from the different transporters.

Received data is 'unpacked' before being delivered to its
trp_client. In order to reduce the context switch overhead
between treads (or trp_clients), we attempt to deliver multiple
signals to the trp_clients before waking them up. This is
implemented as an array of trp_client* which has to
be signaled after the Transporter polling has completed.

If this array is filled to its maks size, we have
to interupt the transporter polling and wakeup the trp_clients
before resuming our activity.

However, when performReceive() resumed, it started over again
on the first Transporter (from the node with the lowest NodeId).
Thus, with sufficient traffic, we could end up in a situation
where performReceive always is interupted before we have received
from all Transporters. This could effectively block all
communication from the higher numbered NodeId

For this particular bug it caused the destruction of
Ndb objects to become stuck as ::close_clnt() required
the (high numbered) API-client to send a signal to itself.
This signal was (almost) never delivered as ::performReceive()
was interupted while receiving from the lower numbered data nodes.