BUG#23197529 REPLICATION IS USING SSL REGARDLESS MASTER_SSL OPTION
Problem:

Slave is always using SSL/TLS when the servers support it, regardless
of MASTER_SSL option.

Analysis:

The client connection on the slave side of the replication must be
configured to use SSL/TLS when MASTER_SSL is set to 1, but it was
doing nothing (using client default configuration) when MASTER_SSL
option was 0.

The issue happened because, since the support for multi-state SSL
option was pushed into 5.7, the default client behavior become to
use SSL, and so, the slave is using SSL if MASTER_SSL = 1 (setting
it up with the replication options) and also if MASTER_SSL = 0 (as
the default behavior is now to use SSL).

Fix:

Setup the slave client connection to not use SSL/TLS when MASTER_SSL
option is set to 0.