BUG#23211701: RESET MASTER must not be possible while Group Replication is running
Executing RESET MASTER while Group Replication is running will purge
all transactions tracking on that member and make member to do not
behave correctly: it will not be able to do more commits since all
will snapshot version will be empty and will be considered a
conflict.

To solve the above issue, we do disallow the execution of RESET
MASTER on Group Replication members.
RESET MASTER is still allowed while Group Replication is installed
but stopped.