Bug#22732697: MYSQL FAILS TO BUILD WITH NEW DEFAULT MODE IN GCC6
GCC 6 has C++14 as default when building C++ code.
Since MySQL does not set any explicit defaults until
the introduction of C++11 with WL#8896, that means that
mysql 5.7 and earlier won't compile with GCC 6.

This patch fixes the problem by explicitly setting
-std=gnu++03 (the previous default) when building with GCC 6+.
No changes when building with earlier versions of GCC or with
other compilers.

This patch is for 5.7.