BUG#23226825: LIMIT NUMBER OF MEMBERS OF GROUP REPLICATION TO 9
Group Replication has the following requirement:
  Group Replication SHALL support up to 9 servers in the group.
  This is obviously dependent on the workload. Write intensive
  workload will REQUIRE high bandwidth networks.

In order to avoid DBAs/users to create groups with a high number of
members, that would reduce the group performance, we are limiting
the group size to 9 members.
This limit does not affect affect connections between groups, or
slaves connected to group members.
This patch introduces a new server error message.