Bug #23068914 DATA NODES CRASHED AT EXECUTEFKCHILDTRIGGER
Foreign keys use triggers to check that referential
integrity is maintained.

Triggers fire when data changes are made at distributed
Local Data Manager(LDM) components, resulting in signals
being sent to the causing transaction's Transaction
Coordinator (TC) component which executes further operations
to ensure integrity.

When a node restarts, these distributed triggers must
be recreated, and it is important that all distributed TC
and LDM component instances agree on trigger identities.

This was not guaranteed by the implementation.

To fix this, an extra step is added to Node Restart, where
trigger identities are determined from the current
running Master node.

A new testcase ndb_fk_restart is added to the ndb MTR suite