Bug #22305994 : innodb_mysql.test fails frequently on PB2 machine due to TIMESTAMP data type is not accurate upto nano seconds on windows
Problem : The innodb_mysql.test is more unstable on PB2 machines due to the reason of timestamp datatype
column precision is not accurate on windows platform , because it is a product bug  on server side.

solution: Remove the failure testcase from innodb.innodb_mysql.test and make it separate as timestamp.test and
disable the testcase (timestamp.test) to run on windows platform, until the actual product issue is resolved.
once the product issue is resolved by dev , he has to enable the testcase (timestamp.test) to run on Windows
This bug will be kept open, until that time (product issue is resolved).

RB : 12561