Bug #22733635   CACHE LINE CONTENTION ON UT_RND_ULINT_COUNTER
PROBLEM

InnoDB implements the random backoff for spinlocks. The function
ut_rnd_interval() is not really
scalable on NUMA core systems since it maintains a global counter
ut_rnd_ulint_counter i.e
updated on each call leading to cache coherency and performance
degradation with many threads
accessing and incrementing the variable.

FIX

As suggested by the bug reporter to make the variable thread local so
that each thread would
maintain a local copy of it and access it reducing the contentions and
delays.
Used TLS library from the server code to make the counter variable
specific to each thread.

Reviewed by: Jimmy Yang<jimmy.yang@oracle.com>
Rb :12444