Bug #22601485: HANDLE_FATAL_SIGNAL (SIG=11) IN PROCESS_BIN_ARG (MEMMOVE)
Problem:
A query containing a format specifier results in server exit when
query rewrite plugin tries to log the query.

Analysis:
When logging the non rewritten query, the text of the query is
passed as the format argument for "my_plugin_log_message".
This makes my_vsnprintf() to treat "%b" in the query as format specifier.
As a result it tries to read the arguments which are not accessible.

Solution:
Pass "%s" as the format argument for "my_plugin_log_message" from
"log_nonrewritten_query" while the query text becomes the argument for
the format.
This fix is a backport from mysql-trunk.