Bug #22661012 : "USING INDEX FOR GROUP-BY" QUERIES CAN RETURN NO    		DATA WITH LE/LEQ AND ROUNDING
Problem:
If rounding happens while storing a predicate value,
range optimizer might not return correct results for < and <= operators.

Analysis:
For the data presented in the bugpage, we have a value for second column
in a multi-part index which is equal to the value that range optimizer
stores after rounding up the predicate value. This makes loose index scan
think it has a qualifying row. But during execution, the row is rejected
because it is larger than the value specified in the where condition.
As a result no rows are selected.

For example:
In the data presented, we have (1,1000), (1,1001), (1,2000) (1,3000)
with 'a' being the first column and 'b' second.

For the query

SELECT a, max(b) FROM t1 WHERE a = 1 and b < 1999.5 GROUP BY a;

we should get row (1,1001).

However since "b" is an int column, range optimizer rounds off 1999.5
to 2000. Loose index scan later finds a row with the matching record
and thinks that this is indeed the max value for the group and
returns (1,2000) as the row which gets rejected later.

The above behavior of loose index scan picking row (1,2000) is
because of the max_flag not being set to NEAR_MAX when range
optimizer detects that the stored value is greater than the
predicate value. In this case, loose index scan needs to start
reading from a value less than the stored value.

However currently, loose index scan starts reading from a value equal
to or less than the value stored.

Solution:
Set NEAR_MAX for max_flag if the stored value is >= than the specified value
for operator < or if the stored value is > than the specified value for the
<= operator.
This patch is contributed by Andrew Bloomgarden.