BUG#22157118:SERVER CRASHES IN MYSQL_BIN_LOG::WRITE_EVENT() WITH MIX/STATEMENT BINLOG-FORMAT
Description:

MySQL Server crashes when 'main.connect' test is executed with
--mysqld=--log-bin and --mysqld=--binlog_format= mixed or statement

Analysis:

If we have a query which uses a user defined variable and after that if
DROP TEMPORARY TABLE is called then the server crashes as the program tries
to read from a corrupted memory part of the  thread variable.

Normally whenever we execute a query it passes through mysql_parse and part of
thread variable which stores information related to user defined variable is
cleaned before doing any sort of processing, but in case of
DROP TEMPORARY TABLE we don't pass through mysql_parse,
so that clean up is not done.

Fix:

Before we call the close_temporary_tables() clear the part of thread variable
which contains information related to user defined variable.