Bug#21609606: BINLOG-FORMAT AND BINLOG_FORMAT SHOW               DIFFERENT BEHAVIOR WHEN PASSED AS CLI OPTION
Issue:
------
Provding the --binlog_format option to mysqld in MTR produces
the correct behaviour, and the same as the server.
However, passing the --binlog-format option to MTR,
causes that option to be removed, if the test does not
have have_log_bin.inc included in it.

Fix:
----
The fix is to make both these options behave correctly.
The following changes can be observed:

1. Currently, passing --binlog_format option without --log-bin
causes the test to fail with a server warning:
"You need to use --log-bin to make --binlog-format work".
After the fix, both --binlog-format and --binlog_format will throw
the same warnings if passed without --log-bin.

2.  Without the patch, if --binlog-format option is passed to a
non-rpl test, with or without --log-bin, it runs with the default
format which is row. It does not run with the specified format.
This patch fixes this makes the test run with the specified format.

3.  Just like before, rpl tests and other tests which include the
have_log_bin.inc file, will run with the specified format of
binlog format, and don't need --log-bin to be specified explicitly.

4. A few tests which do not need to be run with log-bin ON,
but were in the binlog suite, are now moved to the main suite.

5. --log-bin has been added to the runs in the collections
files which have binlog-format.

Reviewed-by: Sayantan Dutta <sayantan.dutta@oracle.com>
Reviewed-by: Venkatesh Duggirala <venkatesh.duggirala@oracle.com>
RB: 10062

Bug#22063875: MYSQLD OPTIONS WITH "-" AND " _ " HANDLED
              DIFFERENTLY BY MTR

Issue:
------
In the mysql-test-run.pl script, many of the command line options
handled contain only '-' in them, and not '_'. This could cause
different behaviour when options with '_' are passed instead of
'_'.

Fix:
----
A function option_equals has been added to MTR, to be used instead of
a direct string comparison of options so that the '-' and '_' in options
are considered to be the same.
Other changes in the script have been done to include regular expressions
which will match both '-' and '_' in command line options while
handling them.

Reviewed-by: Bjorn Munch <bjorn.munch@oracle.com>
Reviewed-by: Sayantan Dutta <sayantan.dutta@oracle.com>
RB: 10770