Bug #23249846: EMPTY SYS DATABASE PREVENTS MYSQL_UPGRADE OF SYS SCHEMA
Problem:

Running mysql_upgrade on a mysqld which has an empty "sys" database
refuses to create the sys schema saying that sys is a user created
database and should not be overwritten.
This is a little too strict and we should allow the sys schema to be
created when an empty sys database already exists.

Fix:

Modified mysql_upgrade to allow successful upgradation even if an
empty sys database already exists i.e. create the necessary database
objects in the existing empty sys database.