Bug#21042813: TESTS IN BINLOG SUITE FAIL WITH VERY LONG PATHS
Issue:
------
When the mysql-test directory is in a very long path of more
than 128 characters, a few binlog suite tests fail due to
result mismatch errors because their load_file
lines get truncated.

Fix:
----
In the fix the select statement is selected AS 'Loaded'
so that a shorter column name is given.

Reviewed-by: Pavan Naik <pavan.naik@oracle.com>
RB: 12536