Bug #11755512: WRONG ERROR REPORTED WHEN ATTEMPTING TO DROP TABLE DURING BACKUP
When a table is altered during an ongoing backup, mysql reports
it as "Unknown error 762". This happens because my_error()
is unable to fetch the correct error message for the error code 762.

Hence, changed the error message being passed to my_error() to
ER_GET_ERRMSG.

The error is now reported as "Got error 762 'Unable to alter table
as backup is in progress' from NDBCLUSTER".