Bug#23202735 CLIENT PERFORMANCE REDUCED DUE TO THREADS WOKEN UP TOO EARLY
This patch fixes two issues related to threads being
woken up when not needed, and thus resulting in unnecessary
context switches.:

1)
Avoid API Clients waiting in the poll queue to be woked up every 10ms.

The 'poll-right' is given to (only!) one of the clients waiting
for do_poll. The rest of the clients just wait there such that
they can be woken up when the poller delivers something to them.
However, a max wait time of 10ms is currently enforced, such that
they may be woken to immediately be put back to sleep. This also
implies dequeuing and requeing in the mutex protected poll queue.

Iff the client threads waiting in the poll queue are properly
signaled when signals are delivered, there should be no need to
wake them up regularly. This patch removes the enforced 10ms max
wait time, and instead wait the time specified by the callee.

2)
Fix a performance problem in ::do_poll() where the polling 'clnt' does not
check whether it has been woken up *itself* before completing the poll.

It is sufficient that only some trp_client's in the poll-queue
received data. ::do_poll will then signal these clients and
give up its poll right, even if the max specified wait_time
have not expired.

Thus, another of the clients in the poll-queue has to be
appointed as the new poller, and a *thread switch* is required
to wake up that thread for more do_poll work.

This patch allows do_poll() to continue polling until
either the max specified wait_time have expired, or
the polling client itself has been woken up (by being
delivered what it waited for) This avoid unnecessary
thread switches between the client threads and thus
reduce the overhead in the API client. This results
in a ~10% performance improvement when the client
threads does the polling themself.

This is similar to the mechanism already implemented
in the receiverThread, which explicit request to 'stay poll owner'
when it has been activated. We still see that having the receiverThread
doing the poll is faster than using the client thread. However,
we believe that most of this is due to the receiver thread allowing
more trp_client to have delivered signals before signaling them
vs the client threads. (256 vs 16)